Updated Ignore Folder Logic
Updated the ignore folder logic, so you just type the
name of the folder as `-i source_code`. When getting
markdown files it will check if folder name is in the path
of the file and if it is it will be ignored. So we don't
end up trying to publish files from `node_modules`.